Generic/DisallowLongArraySyntax: move an intentional parse error test
This commit moves an intentional parse error test to its own file. The
parse error test did not trigger
DisallowLongArraySyntaxSniff::process() as this sniff listen for
T_ARRAY and the `array` keyword without parentheses is tokenized by
PHPCS as T_STRING.

Removing the semicolon fixes that and makes the test effective.