Generic/JumbledIncrementer: fix inconsistency
The name of the sniff uses "Incrementer", while the error message uses "Incrementor". This feels inconsistent.

Based on some light research, both terms can be used interchangable, so I'm proposing to make the error message consistent with the sniff name, as changing the sniff name would be a breaking change.

Includes making the same change in an inline comment.