Generic/DisallowYodaConditions: improve code coverage
Includes removing two unreachable conditions

The first condition is unreachable because PHPCS will never call
DisallowYodaConditionsSniff::process() if there is no non-empty token before a
comparison token. Even if the file contains a syntax error, there must
be at least a PHP opening tag before the comparison token for the method
to be called.

The second condition is unreachable because at this point in the code
there will always be at least two non-empty tokens before the comparison
token. If there is only one, it must be a PHP opening tag and, in this
case, the method will bail before reaching the code that sets the
`$prevIndex` variable.

This commit also documents why a certain line is uncovered by tests - and cannot be
covered, nor can it be removed as external sniffs may call
`DisallowYodaConditionsSniff::isArrayStatic()` directly.