QA/Markdown CS: consistent surrounding pipes for tables
Generally speaking, the "full" table format should always be used - with leading and trailing pipes, but that is a bit awkward for this particular table as a template file like this will typically be edited in a fixed size font, while GH, when using the template to create an issue will use a variable size font.

So, this is a bit of a mix-and-match to, yes, use the full table format, but also make it look reasonable when people edit the table in a new issue.