Generic/InlineControlStructure: fix error when handling `else if`
This commit fixes a bug in this sniff where it would not handle
correctly `else if` statements when there is a comment between the
`else` and `if` tokens.

The affected code was added in
https://github.com/squizlabs/PHP_CodeSniffer/commit/18f98cc9a48bacf026c1a61208d03f6b88c2f69e
to fix a similar bug when there is more than one whitespace between `else`
and `if`, but it failed to consider that there might be comment tokens
between `else` and `if`.