Generic/InlineControlStructure: remove unreachable condition
The removed condition was added in
a5d3f147190dbb2fd103f1163eca69835edafb44. The tests added in this
commit still pass without the removed condition.

The condition is unreachable because `$end` can never be equal to
`$phpcsFile->numTokens`. It will always be less than
`$phpcsFile->numTokens`. `$end` represents the index of a token in the
tokens array. The indexes of the tokens array are zero-based, and the
`$phpcsFile->numTokens` property is one-based. So even if `$end` is
pointing to the index of the last token, it will be less than
`$phpcsFile->numTokens`.

`$end` is first set as `$end = ($closer + 1)`. In theory, if `$closer`
is the index of the last token, `$end` could be equal to
`$phpcsFile->numTokens`. But that is a situation that will never occur
as it will represent a syntax error, and the sniff would have bailed
already.