Generic/InlineControlStructure: remove redundant condition
This commit removes a redundant condition that was added in fbea319 to
support JS braceless do-while loops. A subsequent commit added similar
code to support braceless do-while loops (plus for/while loops without
body) for PHP, but it also works for JS (13c803b).

There are a few syntax error cases that were handled by the code that is
removed by this commit and are not handled by the code introduced in
13c803b. Without the removed code, they are now handled in a if
condition right below. I added two tests with those cases to ensure the
sniff continues working as expected.