Generic/InlineControlStructure: fix bug when handling elseif/if/foreach without body
In PHP, `elseif`, `if` and `foreach` can be declared without a body.
The sniff was improperly flagging those as inline control structures.
This commit changes that, and now the sniff does not trigger an error
anymore for elseif/if/foreach without a body. This is consistent with
the behavior of the sniff for `while` and `for` statements without a
body.

It is likely that this problem was not found before because it is hard
to imagine a case in which an `elseif`, `if` or `foreach` without a body
would be useful.