Generic/InlineControlStructure: remove unreachable condition
This commit removes an unreachable condition in the fixer part of the
sniff. The original version of this condition was added in the early
days by the commit that enabled this sniff to fix errors:

https://github.com/squizlabs/PHP_CodeSniffer/commit/a54c6193472f7c98d64c481a9e7d1203de300889#diff-4b3945c2100b0a92a56509de1b797bf58ad804cf36233c95c492479b665655dcL108-L110

The only two tests that were added with the commit mentioned above
that trigger the removed condition are tests using `while` loops
without body:

https://github.com/squizlabs/PHP_CodeSniffer/commit/a54c6193472f7c98d64c481a9e7d1203de300889#diff-4b3945c2100b0a92a56509de1b797bf58ad804cf36233c95c492479b665655dcL108-L110

I believe control structures without a body are the only cases where
`$next` would be equal to `$end`. Thus, these are the only cases where
the removed condition would be executed. But subsequent commits changed
the sniff to bail early and not get to the fixer part when handling
control structures without a body:

- 13c803b changed the sniff to ignore while/for without a body and
  updated the existing tests: https://github.com/squizlabs/PHP_CodeSniffer/commit/13c803bb14d01641ca089d515eab4287c71e8156#diff-2f069f3fe33bacdfc80485b97303aec66c98c451d07e6d86e41982b81ab1a294L49-R51
- f4afa10 expanded the same approach to also include elseif/if/foreach
  control structures

Those changes rendered the condition that is removed by this commit
unreachable.