Generic/ArrayIndent: avoid extra calls to ArrayIndent::processMultiLineArray()
This commit fixes a small inefficiency when fixing the
`CloseBraceNotNewLine` error. Previously, the sniff would add the newline
and the number of spaces used for an array element (which is more
than what should be used for the array closing brace). Then on a
subsequent call to processMultiLineArray(), it would detect the wrong
number of spaces for the closing brace and it would fix it. Now, the
sniff will use the correct number of spaces when adding the newline.