Tokenizer/PHP: fix handling of "DNF look-a-likes" in named parameters
The last parameter in a function call using named arguments could be confused with a return type by the tokenizer layer handling type declarations.

The net effect of this was that the close parenthesis of the function call would be retokenized to `T_TYPE_CLOSE_PARENTHESIS`, which is incorrect and would lead to sniffs incorrectly acting on that information.

Fixed now.

Includes tests.

Fixes 504
Fixes 505