Generic/DisallowYodaConditions: add code coverage ignore annotation
This commit adds the `@codeCoverageIgnoreStart/@codeCoverageIgnoreEnd`
PHPUnit annotations to ignore, for the purposes of code coverage, a
line that can't be tested. I considered doing this initially when I
worked on improving code coverage for this sniff in #465, but I
wrongly assumed that was not an option due to the
`Squiz.Commenting.InlineComment.InvalidEndChar` error (which is part
of the standard used by PHPCS).

What I failed to notice back then is that I was adding the annotation
below a comment, and this triggers the error:

```
// Shouldn't be possible but may happen if external sniffs are using this method.
// @codeCoverageIgnoreStart
```

Adding the annotation before the comment as it is done in this commit
does not trigger the error because the `Squiz.Commenting.InlineComment`
sniff will only trigger the `InvalidEndChar` error if the first
character of the comment is a letter.

Using `return true; // @codeCoverageIgnore` is not an option because it
triggers `Squiz.Commenting.PostStatementComment.Found`.