Squiz/ArrayDeclaration: improves test coverage
This commit adds tests that were previously missing. There are still
some uncovered lines in this sniff, those lines might be unreachable,
but I did not investigate this at this time. There is also potentially
some more relevant tests that could be added for lines that are already
covered. This is also not addressed in this commit.