Generic/DisallowSpaceIndent: add test with multi-line `yield from`
As the indentation whitespace in a multi-line `yield from` expression is now tokenized as `T_WHITESPACE`, instead of being tokenized as part of the (second) `T_YIELD_FROM` token, space indentation of such a line will now be handled correctly by the sniff.

This commit just adds a test to confirm and safeguard this.